WIP: additional refactor/cleanup from PR review
